Update Gradle version and improve build settings
This commit updates the Gradle version from 8.8 to 8.9 in a few files including the Gradle wrapper properties and the Gradle build files. It also adjusts the settings for the Gradle build, such as updating the Java distribution from Temurin to Liberica in the GitHub workflows, and enhances the script logic in the Gradle wrapper.